// src/app/05_SIR.Carga.Reportes/interfaces/RegistroConPrecio.interface.ts
export interface RegistroConPrecio {
    boxid:          string;            
    clasificacion:  string;
    codigoKosher:   string;
    container:      string;
    expiredate:     string;         
    exportdate:     string;
    finalizada:     boolean;        
    grossweight:    number;       
    id_Carga:       number;           
    id_Pallet:      number;          
    mark:           string;
    netweight:      number;         
    pallet:         number;            
    precio:         number;  
    productcode:    string;        
    productiondate: string;
}