// src/app/05_SIR.Carga.Reportes/interfaces/PLPallet.interface.ts
export interface PLPallet {
    cantidadCajas:          number;
    clasificacionKosher:    string;
    codigoKosher:           string;
    codigoProducto:         string;
    contenedor:             string;
    fechaExpiracion:        string;
    fechaProduccion:        string;
    idPallet:               number;
    markKosher:             string;
    pesoNeto:               number;
    pesoBruto:              number;
    precio:                 number;
}